Adds an update method that allows for update of list title
This method starts by having an edit button which is in the
card component. This edit button takes the user to a route
which is defined in the router file. We then generated a new component
and route both called edit. The edit component has the form in which
you edit the title as well as a cancel button. These two actions then
go to the edit `component.js` file. Those actions are handled by the
app/list/edit folder.